feat(enquete): extract control & content from MenuStepper #1832
